requirements/base.txt
django==4.0.10
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
drf-yasg==1.20.0

Pillow==9.5.0

argon2-cffi==21.3.0

pytz==2023.3